Struct pielectricseagles::config_parse::config_parser

pub struct config_parser {
    config_parser: Ini,
}

It is a config_parser object config_parser class (crate)

Made by: Alex Zaslavskis (Electrics Eagles) 2021

Simple software level that parse own configuration for drone setting that retrieve information about PID-setting, interface and others

Examples

**** Already added to loggics file. Be careful. Editing code can break stability of devices. *****

use crate::config_parse::*;
let mut config = config_parser::new(); // create object of config_parse
let autolevel = config.auto_level_config(); // getting value of auto level
let pid_settings = config.get_pids(); // getting object value of pid-controller of all section
let mpu6050_conifg = config.mpu_config_parser(); // getting object value of configuartion for MPU6050 sensor
let esc_config = config.esc_config_parser(); // getting object value of configuration for esc driver
let version_config: &'static str = config.config_parser_version(); // getting text about version of config parser

Fields
config_parser: Ini
Implementations

impl config_parser

pub fn new() -> config_parser

Returns config_parser object

Arguments

No arguments required

Return

config_parser

Example

use crate::config_parse::*;
let mut config = config_parser::new(); // create object of config_parse

**** Already added to loggics file. Be careful. Editing code can break stability of devices. *****

pub fn parse_u8(&mut self, first_number_str: &str) -> Result<u8, ParseIntError>

Parse i32 value from &str value

Arguments

first_number_str is value with type &str (std:str)

Return

Result<i32, ParseIntError>

Example

**** Already added to loggics file. Be careful. Editing code can break stability of devices. *****

Parse u8 value from &str value

Arguments

first_number_str is value with type &str (std:str)

Return

Result<u8, ParseIntError>

Example

**** Already added to loggics file. Be careful. Editing code can break stability of devices. *****

let mut config = config_parser::new(); // create object of config_parse
let val_u8: u8 = config.parse("55").unwrap(); // getting value u8 from string with numberic value

Parse u32 value from &str value

Arguments

first_number_str is value with type &str (std:str)

Return

Result<u32, ParseIntError>

Example

**** Already added to loggics file. Be careful. Editing code can break stability of devices. *****

let mut config = config_parser::new(); // create object of config_parse
let val_u32: u32 = config.parse("55").unwrap(); // getting value u32 from string with numberic value

Parse f64 value from &str value

Arguments

first_number_str is value with type &str (std:str)

Return

Result<f64, ParseIntError>

Example

let mut config = config_parser::new(); // create object of config_parse
let val_f64: f64 = config.parse("5.51").unwrap(); // getting value f64 from string with numberic value

**** Already added to loggics file. Be careful. Editing code can break stability of devices. *****

Show version of configuration parser

Arguments

No arguments required

Return

&'static str

Example

use crate::config_parse::*;
let mut config = config_parser::new(); // create object of config_parse
let version_config: &'static str = config.config_parser_version();
println!("{}", version_config); // print that shows version into console

**** Already added to loggics file. Be careful. Editing code can break stability of devices. *****

Return data of esc motor configuration from pasring ini-file

Arguments

No arguments required

Return

EscMotors

Parse u8 value from &str value

Arguments

first_number_str is value with type &str (std:str)

Return

Result<u8, ParseIntError>

Example

**** Already added to loggics file. Be careful. Editing code can break stability of devices. *****

let mut config = config_parser::new(); // create object of config_parse
let val_u8: u8 = config.parse("55").unwrap(); // getting value u8 from string with numberic value

pub fn parse_u32(
&mut self,
first_number_str: &str
) -> Result<u32, ParseIntError>

Parse u32 value from &str value

Arguments

first_number_str is value with type &str (std:str)

Return

Result<u32, ParseIntError>

Example

**** Already added to loggics file. Be careful. Editing code can break stability of devices. *****

let mut config = config_parser::new(); // create object of config_parse
let val_u32: u32 = config.parse("55").unwrap(); // getting value u32 from string with numberic value

pub fn parse_f64(
&mut self,
first_number_str: &str
) -> Result<f64, ParseIntError>

Parse f64 value from &str value

Arguments

first_number_str is value with type &str (std:str)

Return

Result<f64, ParseIntError>

Example

let mut config = config_parser::new(); // create object of config_parse
let val_f64: f64 = config.parse("5.51").unwrap(); // getting value f64 from string with numberic value

**** Already added to loggics file. Be careful. Editing code can break stability of devices. *****

pub fn config_parser_version(&mut self) -> &'static str

Show version of configuration parser

Arguments

No arguments required

Return

&'static str

Example

use crate::config_parse::*;
let mut config = config_parser::new(); // create object of config_parse
let version_config: &'static str = config.config_parser_version();
println!("{}", version_config); // print that shows version into console

**** Already added to loggics file. Be careful. Editing code can break stability of devices. *****

pub fn esc_config_parser(&mut self) -> EscMotors

Return data of esc motor configuration from pasring ini-file

Arguments

No arguments required

Return

EscMotors

Example

use crate::config_parse::*;
let mut config = config_parser::new(); // create object of config_parse
let esc_config = config.esc_config_parser(); // getting object value of configuration for esc driver

**** Already added to loggics file. Be careful. Editing code can break stability of devices. *****

pub fn mpu_config_parser(&mut self) -> Mpu6050Config

Return data of MPU6050 gyroscope sensor configuration from pasring ini-file

Arguments

No arguments required

Return

Mpu6050Conifg

Example

**** Already added to loggics file. Be careful. Editing code can break stability of devices. *****

use crate::config_parse::*;
let mut config = config_parser::new(); // create object of config_parse
let mpu6050_conifg = config.mpu_config_parser(); // getting object value of configuartion for MPU6050 sensor

pub fn ibus_receiver_conifg(&mut self) -> SbusConfig

Return data of Sbus/Ibus configuration from pasring ini-file

Arguments

No arguments required

Return

SbusConfig

Example

**** Already added to loggics file. Be careful. Editing code can break stability of devices. *****

use crate::config_parse::*;
let mut config = config_parser::new(); // create object of config_parse
let Sbus = config.ibus_receiver_conifg(); // get object of values S-BUS configuration

pub fn pid_parse(
&mut self,
section: String,
p_str: String,
i_str: String,
d_str: String,
max: String
) -> PID

PID parse for selected section from ini file

Arguments

section - choose section of PID-configarution (String value) p_str - proportional i_str - integral d_str - derivative max - maximun output of PID

Return

PID

Example

**** Already added to loggics file. Be careful. Editing code can break stability of devices. *****

use crate::config_parse::*;
let mut config = config_parser::new(); // create object of config_parse
let pid_value_yaw = config.pid_parse(
       "pid-config".parse().unwrap(),
   "p_pid_yaw".parse().unwrap(),
   "i_pid_yaw".parse().unwrap(),
   "d_pid_yaw".parse().unwrap(),
   "max_yaw".parse().unwrap(),); // get object values of pid of single section from parsing ini log file

pub fn get_pids(&mut self) -> PIDS

Return data of PID-configuration for all type of controls

Arguments

No arguments required

Return

PIDS

Example

**** Already added to loggics file. Be careful. Editing code can break stability of devices. *****

use crate::config_parse::*;
let mut config = config_parser::new(); // create object of config_parse
let pid_settings = config.get_pids(); // getting object values of pid-controller of all section

pub fn auto_level_config(&mut self) -> i32

Return status of auto-level mode

Arguments

No arguments required

Return

i32 0 - fasle 1 - true

Example

**** Already added to loggics file. Be careful. Editing code can break stability of devices. *****

use crate::config_parse::*;
let mut config = config_parser::new(); // create object of config_parse
let autolevel = config.auto_level_config(); // getting value of auto level