Struct pielectricseagles::controller::Controller

pub struct Controller {
    pwm: Pca9685<I2cdev>,
}

It is a Controller object Controller class (crate)

Made by : Misha Zaslavkis (Electrics Eagles) 2021

Simple software level that esc launch for drone. Esc motor lauching with selected pwm and test how rotate motors.

Already added to loggics file. Be careful. Editing code can break stability of devices.

Examples

use crate::controller::Controller;
let mut controller = Controller::new(); // Initialize class of esc driver
controller.calibrate_esc(); // Calibrate esc
controller.set_throttle_external_pwm(1050, 1050, 1050, 1050, 1050); // Run esc's motor with selected pusle width

Fields
pwm: Pca9685<I2cdev>
Implementations

impl Controller

pub fn new() -> Controller

Returns Controller object

Arguments

No arguments required

Examples

***** Already added to loggics file. Be careful. Editing code can break stability of devices. *****

use crate::controller::Controller;
let mut controller = Controller::new();

pub fn set_throttle_external_pwm(
&mut self,
ch1: u16,
ch2: u16,
ch3: u16,
ch4: u16
)

Launch esc motors by selected pwm duty on each motor

Arguments

ch1 - channel 1

ch2 - channel 2

ch3 - channel 3

ch4 - channel 4

All values of channels are range from 1000 - 2000 us

1000 - slowest speed of running esc

2000 - maximun speed of running esc

Examples

**** Already added to loggics file. Be careful. Editing code can break stability of devices. *****

use crate::controller::Controller;
let mut controller = Controller::new(); // create object of Controller
controller.calibrate_esc(); // Calibrate esc. This function is mandatory to use before starting run esc
controller.set_throttle_external_pwm(1050, 1050, 1050, 1050, 1050); // Set specifity pusle of pwm generate esc driver

pub fn test_esc(&mut self, pwm_duty: u16)

Testing launching esc motors by selected pwm duty for all motor.

Value of pmw duty is same for all motors

To avoid doesn’t work esc, use function of calibarating esc. Call function calibrate_esc(); before launch esc.

Arguments

pwm_duty - speed to test how launch esc motors

1000 - slowest speed of running esc

2000 - maximun speed of running esc

Examples

**** Already added to loggics file. Be careful. Editing code can break stability of devices. *****

use crate::controller::Controller;
let mut controller = Controller::new();
controller.calibrate_esc();
controller.test_esc(1500);

pub fn calibrate_esc(&mut self)

Calibration all esc motors

This fuction give teachering the esc controller what range of output throttle input it should respond to. A function give range of pulse 1000-2000 to teaching the esc controller. Before usage esc, you must use this function to works property esc. This function is mandatory!

Arguments

No arguments required

Examples

**** Already added to loggics file. Be careful. Editing code can break stability of devices. *****

use crate::controller::Controller;
let mut controller = Controller::new();
controller.calibrate_esc();